Add DES and 3DES encryption and decryption
Add BlockOperation trait which represents a cryptographic operation
on a single block of input. Implement this trait for performing DES
and 3DES encryption and decryption on a block of input.

Add padding module which contain Read implementations for adding
and remove PKCS5 padding from an underlying reader.

Implement CBC mode for encryption and decryption operations.

The choice of block encryption/decryption mode, block-chaining
mode and padding scheme allows encryption and decryption to be
performed generically. Add DES/3DES encrypt/decrypt functions for
encrypting a buffer of plaintext or ciphertext data.

Add tests for these basic operations for the examples in the book.

Add des program which takes the operation to perform (encryption
or decryption), the initial IV and key along with the plaintext
or ciphertext to encrypt or decrypt. This performs DES or 3DES
encryption depending on the key length (3DES requires 24-byte keys)
and writes the result encoded as hex to stdout.