Add mode argument to des command
Add mode argument which specifies whether to encrypt or decrypt
the input using the key and IV. Add read_bytes function which
converts a command-line string into a byte vector. If the
argument begins with 0x it is assumed to be a hex-encoded string
otherwise the underlying bytes are used directly.

Move removal of padding within des_decrypt into its own function.