Implement DES3
Parameterise the encryption and decryption process by an
implementation of new BlockOperation trait. This performs a single
operation on a block on input within the encryption or decryption
processes. Add implementations for the encryption and decryption
processes for DES and DES3. Use these to implement the DES3
encrypt and decryption processes using the existing implementation.

Support DES3 within the des program. DES3 is used within the
encryption and decryption modes if the key length is 24, otherwise
DES is used.