Implement RSA encryption and decryption
Add Huge type represented as a big-endian vector of bytes. This type
supports add, subtract, multiply, divide and power operations, along
with the PartialEq, Eq, PartialOrd and Ord traits. It also supports
an efficient 'power modulo n' operation which calculates the
remainder modulo n of self^e. This operation is used extensively in
the RSA encryption and decryption operations.

Implement From and TryFrom implementations between Huge and u8 and
usize.

Add rsa module containing rsa_compute function for encrypting or
decrypting using RSA. Use this to implement rsa_encrypt and
rsa_decrypt functions. These add or remove the required padding to
construct a block to be encrypted or to recover the decrypted
plaintext. Both accumulate the ciphertext or plaintext into a
Vec buffer.

Add rsa program which encrypts or decrypts the given input using RSA
with either the specified key or a hard-coded test key.

Add tests for encryption and decryption using the book example.