Implement MD5 hash
Add two implementations of the MD5 block operation from the book.
Add md5_hash function which returns the internal state of the
hash buffer (4 32-bit unsigned ints) along with and md5_hash_bytes
function for converting this state into an array of bytes.

Add tests for hashing the empty string and the bytes for the 'abc'
string as given in the book.

Add basic digest utility for hashing input strings with md5.