Implement SHA1 hashing
Add sha module containing a sha1 implementation. The sha1_hash
function converts an input byte slice into a final SHA1 state
containing 5 4-byte integers. This can be converted into an
output byte vector with the sha1_hash_bytes function.

Add tests for hashing the empty string and the string 'abc' from
the book.

Add support for SHA1 to the digest utilty and add a command-line
parameter for the digest function to use - MD5 or SHA1.