Make Huge numbers signed
Add sign flag to Huge which indicates whether the value is negative.
Update the add, subtract, multiply and divide functions to take
account of the sign of the arguments and set the sign of the result
accordingly. Remainders of the divide operation can be negative.

Add conversions to and from i8 and isize to allow negative Huge
instances to be constructed directly.