Allow hash values to be calculated incrementally
Add HashState struct which wraps a Vec<u32> to store hash states.
This is compatible with all existing hash algorithms.

Remove the State type from the HashAlgorithm trait and update
methods to create and update instances of HashState instead.

Add BlockBuffer which incrementally copies data into a fixed buffer
and notifies the caller when the buffer has been filled.

Add Digest<H> struct which contains a block buffer, hash algorithm
and hash state which is incrementally updated as data blocks are
written to it.

Add utility digest_hash function similar to the existing hash
function which uses a Digest.