Split up tls module
Convert tls module into directory and add serde, secure, messages
and protocol child modules.

The serde module contains the low-level binary serialisation traits
and associated utilty functions and definitions for built-in types.

The secure module contains definitions for the CipherSuite and
ProtectionParameters types used to secure the connection.

The messages module contains defintions for messages sent to and
received from the server.

The protocol module implements the TLS protocol.